+++ lib/l10n/app_localizations.dart
@@ -1366,12 +1366,6 @@ abstract class AppLocalizations {
   /// **'Privacy policy'**
   String get settingsPrivacyPolicy;
 
-  /// No description provided for @settingsAdPrivacyOptions.
-  ///
-  /// In en, this message translates to:
-  /// **'Ad privacy settings'**
-  String get settingsAdPrivacyOptions;
-
   /// No description provided for @settingsLegalLinkFailed.
   ///
   /// In en, this message translates to: